+++ app/Http/Controllers/WorkoutSectionController.php
@@ -3,12 +3,15 @@
 namespace App\Http\Controllers;
 
 use App\Contracts\Repositories\Repository;
+use App\Events\WorkoutAdding;
 use App\Models\{WorkoutSection,WorkoutImage,Workout};
  
 use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
 use Illuminate\Contracts\Filesystem\Filesystem;
 use Illuminate\Queue\Console\WorkCommand;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Storage;
 
 class WorkoutSectionController extends Controller
@@ -32,8 +35,10 @@ class WorkoutSectionController extends Controller
         $sections = $this->repository->all()->where("workout_section_id",null)->all();
                   
         foreach($sections as $key=>$section){
-            if($section->image)
-                $sections[$key]["image"] =  \Thumbnail::src( env( 'APP_URL' ) .  $section->image->path )->smartcrop(300, 220)->url( true );    
+          
+             if($section->image)
+                $sections[$key]["image"] =  \Thumbnail::src( env( 'APP_URL' ) .'/storage/'.  $section->image->path )->smartcrop(300, 220)->url( true );    
+                
         } 
          
         return view("workout.index",["sections"=>$sections]);
@@ -47,7 +52,7 @@ class WorkoutSectionController extends Controller
     public function create()
     {
        
-        $sections = \App\Models\WorkoutSection::all("id","title");
+        $sections = $this->repository->all();
         $sectAll = [0 => null];
         foreach ( $sections as $value) {
             $sectAll[$value["id"]] = $value["title"];
@@ -70,25 +75,25 @@ class WorkoutSectionController extends Controller
        
         $request->validate($validate_rules);
         
-        $fields = $request->only("title","slug","description","workout_section_id","file");
+        $fields = $request->only("title","slug","description" ,"file");
       
-         
+        
         $workoutSection = WorkoutSection::create([
             "title" => $fields["title"],
             "slug"  => ($fields["slug"]) 
                 ? \Illuminate\Support\Str::slug($fields["slug"],"_")
                 :\Illuminate\Support\Str::slug($fields["title"],"_"),
             "description"   =>  $fields["description"],
-            "workout_section_id"=>($fields["workout_section_id"]) ? $fields["workout_section_id"] : null
+            
+          //  "workout_section_id"=>($fields["workout_section_id"]) ? $fields["workout_section_id"] : null
         ]);
 
-
-        //$workoutSection->save();
+ 
  
         if($request->hasFile("file")){
             WorkoutImage::uploadImage($request->file("file"), $workoutSection);
         }
-        
+
         return redirect()->route("workout.index");
     }
 
@@ -100,13 +105,14 @@ class WorkoutSectionController extends Controller
      */
     public function show(WorkoutSection $workoutSection)
     {
-    
+        $el = resolve(Repository::class);
+      
         //$sections = WorkoutSection::where("workout_section_id",$workoutSection->id)->get();
         $sections = $workoutSection->sections;
                   
         foreach($sections as $key=>$section){
             if($section->image)
-                $sections[$key]["image"] =  \Thumbnail::src( env( 'APP_URL' ) .  $section->image->path )->smartcrop(300, 220)->url( true );    
+                $sections[$key]["image"] =  \Thumbnail::src( env( 'APP_URL' )  .'/storage/'. $section->image->path )->smartcrop(300, 220)->url( true );    
         } 
 
         //$workouts = Workout::where(["workout_section_id"=>$workoutSection->id])->get();   
@@ -114,7 +120,7 @@ class WorkoutSectionController extends Controller
 
         foreach($workouts as $key=>$workout){
             if($workout->image)
-                $workouts[$key]["image"] =  \Thumbnail::src( env( 'APP_URL' ) .  $workout->image->path )->smartcrop(300, 220)->url( true );    
+                $workouts[$key]["image"] =  \Thumbnail::src( env( 'APP_URL' )  .'/storage/'.  $workout->image->path )->smartcrop(300, 220)->url( true );    
         } 
         return view("workout.section.index",["section"=>$workoutSection,"workouts"=>$workouts,"sections"=>$sections]);
     }
@@ -153,7 +159,7 @@ class WorkoutSectionController extends Controller
 
         $request->validate($validate_rules);
 
-        $fields = $request->only(["title","slug","description","workout_section_id","file"]); 
+        $fields = $request->only(["title","slug","description","file"]); 
 
         $workoutSection->update([
             "title" => $fields["title"],
@@ -161,13 +167,15 @@ class WorkoutSectionController extends Controller
                 ? \Illuminate\Support\Str::slug($fields["slug"],"_")
                 :\Illuminate\Support\Str::slug($fields["title"],"_"),
             "description"   =>  $fields["description"],
-            "workout_section_id"=> ($fields["workout_section_id"]) ? $fields["workout_section_id"] : null
+            //"workout_section_id"=> ($fields["workout_section_id"]) ? $fields["workout_section_id"] : null
         ]);
 
         if($request->hasFile("file")){
            
-            Storage::delete($workoutSection->image->path);
-            $workoutSection->image->delete();
+            if($workoutSection->image){
+                Storage::delete($workoutSection->image->path);
+                $workoutSection->image->delete();
+            }
 
             WorkoutImage::uploadImage($request->file("file"), $workoutSection);
